Adding full example for product-resources
It took me a lot of googling to find a mention of the full syntax for instance_type and persistent_disk for this section.   Strange the names of the parameters differ so much from the source of the web ui.  I didn't look at a metadatafile though, hopefully it matches there?