Add `certificate-authority` command
This change adds a new `certificate-authority` command that will
retrieve a specific certificate authority identified by its ID.

This change also renames the certificateAuthorityLister to
certificateAuthoritiesService to be more consistent with existing
interface names.

This change also moves the presenter fake to be local to where the
interface is defined.

[finishes #152444865]

Signed-off-by: James Myers <jmyers@pivotal.io>