support -c flag for configure-director
The configure-director command takes quite a few arguments that are JSON strings.
This can make it difficult to use from the command line.
The --config/-c option allows an equivalant YAML file to be used in place of the other flags.
It will return the same errors message as if the flags were invoked directly.