staged-director-config uses new `properties-configuration` field [#162794545]
This changes the config file exported to have the fields of `director-configuration`, `iaas-configuration`, `syslog-configuration`, and `security-configuration` to match under the top level key `properties-configuration`.
That field is the wrappers for the director properties endpoint.
This means as new fields are added to the endpoint they will exported without requiring future changes to `om` to support them.

#310