use reflection rather than type assertion when pruning the director config [Fixes #372]
The issue with type assertion was that there were too many types to assert about.
For example, a map of maps or a slice of maps.
The reflection usages just makes the type care that it is a map or a slice that has values, not caring what that value is.