bosh-env command does not resolve private key
Right now, `om bosh-env -i keyname` does no checking if a private key
is valid or not. That, in and of itself, is not a big deal, but it also
doesn't create an absolute path to the key. This PR addresses that. When
specifying a key via `-i` it will attempt to open the file, and if it
is successful, will use its absolute path in `BOSH_ALL_PROXY` and
`CREDHUB_PROXY`. If the file is not openable, it uses just the provided
string, to provide backwards behavior compatibility.