remove --skip-unchanged-products
This option has had issues with consistent successful behaviour.
For example, if the apply changes fails for any reason, the subsequent apply changes cannot pick where it left off.
This usually happens in the case of errands that are used for services.

We are working on scoping a selective deploy feature that makes sense for users.
We would love to have feedback from users about this.

Signed-off-by: JT Archie <jarchie@pivotal.io>
Signed-off-by: Kira Boyle <kboyle@pivotal.io>