Configure-director now supports multiple iaases
on iaas (vsphere and openstack) that support them.

The config file now takes in a 'iaas_configuration_name'
instead of 'iaas_configuration_guid' for az configuration.

On multi-iaas configs, this means that the config file
can be defined once. Previously, the guid was required
and was only available after configuration.

[#166483367]

Signed-off-by: JT Archie <jarchie@pivotal.io>