Allow alternatives for config-template cli args
This resolves #432. Because `download-product` can accept required
arguments in various ways (on the command line, via vars file, vars,
or env variables) and because `config-template` has as its options
a subset of `download-product`'s argumennts, it makes sense to allow
`config-template` to accept arguments in the same fashion. This PR
refactors `config-template` to use the same common code as
`download-product` and includes the same tests as well. In addition,
`commands.NewConfigTemplate` needs to accept any `environFunc` instead
of hardcoded `os.Environ`, so a second `New*` method was introduced to
ensure backward compatibility.