use bosh parsing for vars, vars env, and vars file
When using interpolation, users would have an inconsistent experience with bosh and om.
The behaviour was because of our large custom interpolation command.
The bosh CLI code has a lot of error handling and integrity checking, which was duplicated here.

That code has been removed.
Using the bosh CLI code simplifies our processing of vars.
The logic we have for multiline strings had to be readded.
[#171702208]