pulled in vm-lifecycle commands from p-automator
This allows the `om` CLI to handle the creation of the create and delete of an OpsManager VM.
This works allows a single binary to manage your OpsManager.
The requirement, though, for each IAAS that will be interacted with will need the corresponding IAAS CLI.
For example, AWS needs `aws`, vSphere needs `govc`, etc.

[#174956239]

Signed-off-by: Kira Boyle <kboyle@vmware.com>
Signed-off-by: JT Archie <jtarchie@gmail.com>
Signed-off-by: Kira Boyle <kboyle@vmware.com>