Allow to specify separate UAA Host for local dev
allow to reset uaa host to value from env var. in local dev envs we
do not always have nginx or another lb to do path based routing and
therefore can run UAA on a different port. This is not exposed as a
flag to avoid customer iritation about the possibility of resetting
the UAA Host.