fix: Ensure image has go 1.22.2
I couldn't figure out an easier way to update the goreleaser/goreleaser
image.  If I had my 'druthers, I'd have created a new image "from"
goreleaser and add a step to update go.  Making the task do this work
seems the most expedient, but it's a kludge, so you'll want to revert
this in favor of a better solution in the future.

Co-authored-by: Steve Taylor <sttaylor@vmware.com>